Move things around to structure things a bit more cleanly
