+++ src/Components/css/Details.css
